Move barcode.dart
