+++ lib/barcode/barcode.dart
