+++ lib/widget/location_display.dart
@@ -4,7 +4,7 @@ import "package:flutter_speed_dial/flutter_speed_dial.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
 
 import "package:inventree/app_colors.dart";
-import "package:inventree/barcode.dart";
+import "package:inventree/barcode/barcode.dart";
 import "package:inventree/l10.dart";
 
 import "package:inventree/inventree/stock.dart";
